# docs/swagger.yaml
basePath: /api
definitions:
...
  storage.Trigger:
    properties:
      active:
        type: boolean
      config:
        additionalProperties: true
        type: object
      created_at:
        type: string
      error_message:
        type: string
      id:
        type: integer
      last_execution:
        type: string
      name:
        type: string
      next_execution:
        type: string
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Enterprise-grade webhook routing and data transformation platform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Webhook Router API
  version: "1.0"
paths:
...
  /triggers:
    get:
      description: Returns a list of all configured triggers with optional filtering
      parameters:
      - description: Filter by trigger type
        in: query
        name: type
        type: string
      - description: Filter by trigger status
        in: query
        name: status
        type: string
      - description: Filter by active status
        in: query
        name: active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: List of triggers
          schema:
            items:
              $ref: '#/definitions/storage.Trigger'
            type: array
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - SessionAuth: []
      summary: Get all triggers
      tags:
      - triggers
    post:
      consumes:
      - application/json
      description: Creates a new trigger configuration for automated webhook processing
      parameters:
      - description: Trigger configuration
        in: body
        name: trigger
        required: true
        schema:
          $ref: '#/definitions/storage.Trigger'
      produces:
      - application/json
      responses:
        "201":
          description: Created trigger
          schema:
            $ref: '#/definitions/storage.Trigger'
        "400":
          description: Invalid JSON
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - SessionAuth: []
      summary: Create trigger
      tags:
      - triggers
  /triggers/{id}:
    delete:
      description: Removes a trigger configuration
      parameters:
      - description: Trigger ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid trigger ID
          schema:
            type: string
        "500":
          description: Failed to delete trigger
          schema:
            type: string
      security:
      - SessionAuth: []
      summary: Delete trigger
      tags:
      - triggers
    get:
      description: Returns a specific trigger configuration by ID
      parameters:
      - description: Trigger ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Trigger configuration
          schema:
            $ref: '#/definitions/storage.Trigger'
        "400":
          description: Invalid trigger ID
          schema:
            type: string
        "404":
          description: Trigger not found
          schema:
            type: string
      security:
      - SessionAuth: []
      summary: Get trigger
      tags:
      - triggers
    put:
      consumes:
      - application/json
      description: Updates an existing trigger configuration
      parameters:
      - description: Trigger ID
        in: path
        name: id
        required: true
        type: integer
      - description: Trigger configuration
        in: body
        name: trigger
        required: true
        schema:
          $ref: '#/definitions/storage.Trigger'
      produces:
      - application/json
      responses:
        "200":
          description: Updated trigger
          schema:
            $ref: '#/definitions/storage.Trigger'
        "400":
          description: Invalid JSON or trigger ID
          schema:
            type: string
        "500":
          description: Failed to update trigger
          schema:
            type: string
      security:
      - SessionAuth: []
      summary: Update trigger
      tags:
      - triggers
  /triggers/{id}/start:
    post:
      description: Starts a specific trigger
      parameters:
      - description: Trigger ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Start result
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid trigger ID
          schema:
            type: string
        "404":
          description: Trigger not found
          schema:
            type: string
        "500":
          description: Failed to start trigger
          schema:
            type: string
      security:
      - SessionAuth: []
      summary: Start trigger
      tags:
      - triggers
  /triggers/{id}/stop:
    post:
      description: Stops a specific trigger
      parameters:
      - description: Trigger ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Stop result
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid trigger ID
          schema:
            type: string
        "404":
          description: Trigger not found
          schema:
            type: string
        "500":
          description: Failed to stop trigger
          schema:
            type: string
      security:
      - SessionAuth: []
      summary: Stop trigger
      tags:
      - triggers
  /triggers/{id}/test:
    post:
      description: Tests a trigger configuration
      parameters:
      - description: Trigger ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Test result
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid trigger ID or configuration
          schema:
            type: string
        "404":
          description: Trigger not found
          schema:
            type: string
      security:
      - SessionAuth: []
      summary: Test trigger
      tags:
      - triggers
  /triggers/types:
    get:
      description: Returns a list of supported trigger types
      produces:
      - application/json
      responses:
        "200":
          description: List of trigger types
          schema:
            additionalProperties: true
            type: object
      security:
      - SessionAuth: []
      summary: Get trigger types
      tags:
      - triggers
...
securityDefinitions:
  SessionAuth:
    in: cookie
    name: session
    type: apiKey
swagger: "2.0"